backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
asyncpg==0.29.0
pdfplumber==0.11.0
python-docx==1.1.2
httpx==0.27.0
python-multipart==0.0.9
pydantic==2.7.4
